Refactored utils to use new MVP modules. Closes #30
